+++ modules/validators/src/include/Validators.hpp
@@ -81,5 +81,8 @@ mustBeNonnegative(const ArrayOfVector& args, int argPosition);
 NLSVALIDATORS_IMPEXP void
 mustBeNegative(const ArrayOfVector& args, int argPosition);
 //=============================================================================
+NLSVALIDATORS_IMPEXP void
+mustBeNonNan(const ArrayOfVector& args, int argPosition);
+//=============================================================================
 }
 //=============================================================================